+++ types/milesight/EM400-MUD/uplink.spec.js
@@ -1,0 +1,67 @@
+const chai = require("chai");
+
+const rewire = require("rewire");
+const utils = require("test-utils");
+
+const { assert } = chai;
+
+describe("EM400-MUD Uplink", () => {
+  let defaultSchema = null;
+  let lifecycleSchema = null;
+  let consume = null;
+  before((done) => {
+    const script = rewire("./uplink.js");
+    consume = utils.init(script);
+    utils
+      .loadSchema(`${__dirname}/default.schema.json`)
+      .then((parsedSchema) => {
+        defaultSchema = parsedSchema;
+        done();
+      });
+  });
+
+  before((done) => {
+    const script = rewire("./uplink.js");
+    consume = utils.init(script);
+    utils
+      .loadSchema(`${__dirname}/lifecycle.schema.json`)
+      .then((parsedSchema) => {
+        lifecycleSchema = parsedSchema;
+        done();
+      });
+  });
+
+  describe("consume()", () => {
+    it("should decode should decode the EM400-MUD payload", () => {
+      const data = {
+        data: {
+          port: 1,
+          payloadHex: "01755C0367010104824408050001",
+        },
+      };
+
+      utils.expectEmits((type, value) => {
+        assert.equal(type, "sample");
+        assert.isNotNull(value);
+        assert.typeOf(value.data, "object");
+
+        assert.equal(value.data.batteryLevel, 92);
+        utils.validateSchema(value.data, lifecycleSchema, { throwError: true });
+      });
+
+      utils.expectEmits((type, value) => {
+        assert.equal(type, "sample");
+        assert.isNotNull(value);
+        assert.typeOf(value.data, "object");
+
+        assert.equal(value.data.temperature, 25.7);
+        assert.equal(value.data.distance, 2116);
+        assert.equal(value.data.tilt, true);
+
+        utils.validateSchema(value.data, defaultSchema, { throwError: true });
+      });
+
+      consume(data);
+    });
+  });
+});